flashcard-generator/requirements.txt
streamlit>=1.32.0
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.30.0
pillow>=10.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
pytesseract>=0.3.10
gTTS>=2.3.0
pydub>=0.25.0
ffmpeg-python>=0.2.0
fpdf>=1.7.0
sentence-transformers>=2.2.0
numpy>=1.26.0
pandas>=2.2.0
requests>=2.31.0
python-dateutil>=2.9.0
tqdm>=4.66.0
protobuf>=4.25.0
safetensors>=0.4.0
tokenizers>=0.15.0
huggingface-hub>=0.22.0
watchdog>=4.0.0
# Using a more stable translation library
deep-translator>=1.11.0
